refactor(web): split UnifiIntegration.tsx into per-card components + state hook (#2382)
UnifiIntegration.tsx was ~2,314 lines — one stateful component holding the
entire UniFi integration surface. Split into the seams the issue proposed,
with zero behavior/API/copy/i18n-key change:

- useUnifiIntegration() — owns all status/hosts/orgs/sites/collectors/
  controllers/telemetry state plus the fetch/handlers.
- ConnectionChooser — disconnected screen (cloud vs self-hosted sub-forms).
- ConnectedSummaryCard — account/last-sync, reauth/error banners, sync/
  disconnect actions.
- ControllersCard — self-hosted controller registration + list.
- SiteMappingCard — the two near-duplicate cloud/self-hosted mapping blocks,
  kept as one component with a `variant` discriminant so each JSX branch
  stays a verbatim move rather than a risky merge.
- TelemetryCollectorsCard + TelemetryViewerCard — deep-telemetry collector
  config and the live telemetry viewer (cloud).
- SyncHistoryCard — sync-run ledger (cloud).
- unifiTypes.ts / unifiHelpers.ts — shared types and pure helpers
  (agentLabel, agentOptionLabel, loadAgentDevices, status→badge-class maps).

UnifiIntegration.tsx now just calls the hook and wires the cards together
(2,314 → 274 lines). UnifiIntegration.test.tsx is unchanged and green (23/23,
data-testid based) since every data-testid, string, and DOM structure moved
verbatim.
